+++ src/main/java/kmu/maplayers/politicalmap/dominance/alliances/AlliancesView.java
@@ -3,7 +3,7 @@ package kmu.maplayers.politicalmap.dominance.alliances;
 import com.fs.starfarer.api.campaign.SectorAPI;
 import com.fs.starfarer.api.impl.campaign.ids.Factions;
 
-import kmlib.starsector.ui.controls.ControlSpec;
+import kmlib.starsector.ui.controls.specs.ControlSpec;
 
 import kmu.maplayers.base.refresh.MapLayerRefreshBoard;
 import kmu.maplayers.base.theme.ElementStyleAdjustment;