+++ src/test/java/kmu/maplayers/base/sidebar/BarOpenersTest.java
@@ -1,6 +1,6 @@
 package kmu.maplayers.base.sidebar;
 
-import kmlib.starsector.ui.controls.ControlSpec;
+import kmlib.starsector.ui.controls.specs.ControlSpec;
 import kmlib.starsector.ui.widgets.tabs.BandButtonSpec;
 import kmlib.starsector.ui.widgets.tabs.style.TabBox;