+++ src/test/java/kmu/maplayers/politicalmap/dominance/alliances/AlliancesViewTest.java
@@ -7,7 +7,7 @@ import com.fs.starfarer.api.campaign.FactionAPI;
 import com.fs.starfarer.api.campaign.SectorAPI;
 import com.fs.starfarer.api.impl.campaign.ids.Factions;
 
-import kmlib.starsector.ui.controls.ControlSpec;
+import kmlib.starsector.ui.controls.specs.ControlSpec;
 import kmlib.starsector.ui.widgets.lists.ListSortMode;
 
 import kmu.maplayers.base.layer.ScreenMemoryScopes;